Migrate StatusIconAndText
