+++ frontend/packages/console-dynamic-plugin-sdk/src/app/components/index.ts
@@ -1,3 +1,6 @@
 export { default as GenericStatus } from './status/GenericStatus';
 export { default as PopoverStatus } from './status/PopoverStatus';
+export { default as StatusIconAndText } from './status/StatusIconAndText';
 export { default as ResourceStatus } from './utils/resource-status';
+
+export * from './status/icons';
