+++ frontend/packages/console-shared/src/components/status/LinkStatus.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as History from 'history';
 import { Link } from 'react-router-dom';
-import StatusIconAndText from './StatusIconAndText';
+import { StatusIconAndText } from '@console/dynamic-plugin-sdk';
 
 const LinkStatus: React.FC<LinkStatusProps> = ({ linkTitle, linkTo, ...other }) =>
   linkTo ? (
