+++ frontend/packages/console-shared/src/components/status/icons.tsx
@@ -1,10 +1,6 @@
 import * as React from 'react';
 import {
   ArrowCircleUpIcon,
-  CheckCircleIcon,
-  ExclamationCircleIcon,
-  ExclamationTriangleIcon,
-  InfoCircleIcon,
   UnknownIcon,
   SyncAltIcon,
   ResourcesAlmostFullIcon,
@@ -14,48 +10,16 @@ import { global_danger_color_100 as dangerColor } from '@patternfly/react-tokens
 import { global_default_color_200 as blueDefaultColor } from '@patternfly/react-tokens/dist/js/global_default_color_200';
 import { global_disabled_color_100 as disabledColor } from '@patternfly/react-tokens/dist/js/global_disabled_color_100';
 import { global_palette_blue_300 as blueInfoColor } from '@patternfly/react-tokens/dist/js/global_palette_blue_300';
-import { global_palette_green_500 as okColor } from '@patternfly/react-tokens/dist/js/global_palette_green_500';
 import { global_warning_color_100 as warningColor } from '@patternfly/react-tokens/dist/js/global_warning_color_100';
+import { ColoredIconProps } from '@console/dynamic-plugin-sdk';
 
-export const GreenCheckCircleIcon: React.FC<ColoredIconProps> = ({ className, title, size }) => (
-  <CheckCircleIcon
-    data-test="success-icon"
-    size={size}
-    color={okColor.value}
-    className={className}
-    title={title}
-  />
-);
-
-export const RedExclamationCircleIcon: React.FC<ColoredIconProps> = ({
-  className,
-  title,
-  size,
-}) => (
-  <ExclamationCircleIcon
-    size={size}
-    color={dangerColor.value}
-    className={className}
-    title={title}
-  />
-);
-
-export const YellowExclamationTriangleIcon: React.FC<ColoredIconProps> = ({
-  className,
-  title,
-  size,
-}) => (
-  <ExclamationTriangleIcon
-    size={size}
-    color={warningColor.value}
-    className={className}
-    title={title}
-  />
-);
-
-export const BlueInfoCircleIcon: React.FC<ColoredIconProps> = ({ className, title }) => (
-  <InfoCircleIcon color={blueInfoColor.value} className={className} title={title} />
-);
+export {
+  ColoredIconProps,
+  GreenCheckCircleIcon,
+  RedExclamationCircleIcon,
+  YellowExclamationTriangleIcon,
+  BlueInfoCircleIcon,
+} from '@console/dynamic-plugin-sdk';
 
 export const GrayUnknownIcon: React.FC<ColoredIconProps> = ({ className, title }) => (
   <UnknownIcon color={disabledColor.value} className={className} title={title} />
@@ -76,9 +40,3 @@ export const YellowResourcesAlmostFullIcon: React.FC<ColoredIconProps> = ({ clas
 export const BlueArrowCircleUpIcon: React.FC<ColoredIconProps> = ({ className, title }) => (
   <ArrowCircleUpIcon color={blueDefaultColor.value} className={className} title={title} />
 );
-
-export type ColoredIconProps = {
-  className?: string;
-  title?: string;
-  size?: 'sm' | 'md' | 'lg' | 'xl';
-};
