+++ frontend/packages/console-shared/src/components/status/index.tsx
@@ -4,5 +4,4 @@ export * from './Status';
 
 export { default as SecondaryStatus } from './SecondaryStatus';
 export { default as LinkStatus } from './LinkStatus';
-export { default as StatusIconAndText } from './StatusIconAndText';
 export { default as NodeUnschedulableStatus } from './NodeUnschedulableStatus';
