+++ frontend/packages/kubevirt-plugin/src/components/create-vm-wizard/tabs/import-providers-tab/providers/vm-import-provider-controller-status-row.tsx
@@ -6,10 +6,10 @@ import * as _ from 'lodash';
 import { Trans, useTranslation } from 'react-i18next';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
+import { StatusIconAndText } from '@console/dynamic-plugin-sdk';
 import { resourcePath } from '@console/internal/components/utils';
 import { DeploymentModel, PodModel } from '@console/internal/models';
 import { K8sResourceKind, PodKind } from '@console/internal/module/k8s';
-import StatusIconAndText from '@console/shared/src/components/status/StatusIconAndText';
 import { getName, getNamespace } from '../../../../../selectors';
 import { PodDeploymentStatus } from '../../../../../statuses/pod-deployment/constants';
 import { getPodDeploymentStatus } from '../../../../../statuses/pod-deployment/pod-deployment-status';
