+++ frontend/packages/kubevirt-plugin/src/components/create-vm-wizard/tabs/import-providers-tab/providers/vm-import-provider-object-status.tsx
@@ -3,7 +3,7 @@ import { Alert, AlertVariant } from '@patternfly/react-core';
 import { InProgressIcon } from '@patternfly/react-icons';
 import { useTranslation } from 'react-i18next';
 import { connect } from 'react-redux';
-import StatusIconAndText from '@console/shared/src/components/status/StatusIconAndText';
+import { StatusIconAndText } from '@console/dynamic-plugin-sdk';
 import { V2VProviderStatus } from '../../../../../statuses/v2v';
 import { iGet } from '../../../../../utils/immutable';
 import { FormFieldRow } from '../../../form/form-field-row';
