+++ frontend/packages/operator-lifecycle-manager/src/components/operator-install-page.tsx
@@ -11,7 +11,7 @@ import {
 import { Helmet } from 'react-helmet';
 import { useTranslation } from 'react-i18next';
 import { Link } from 'react-router-dom';
-import { ResourceStatus } from '@console/dynamic-plugin-sdk';
+import { ResourceStatus, StatusIconAndText } from '@console/dynamic-plugin-sdk';
 import { SyncMarkdownView } from '@console/internal/components/markdown-view';
 import { errorModal } from '@console/internal/components/modals';
 import {
@@ -28,7 +28,7 @@ import {
   referenceFor,
   K8sResourceKind,
 } from '@console/internal/module/k8s';
-import { parseJSONAnnotation, StatusIconAndText } from '@console/shared';
+import { parseJSONAnnotation } from '@console/shared';
 import {
   GreenCheckCircleIcon,
   RedExclamationCircleIcon,
