+++ frontend/public/components/overview/build-overview.tsx
@@ -6,7 +6,8 @@ import * as React from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import { SyncAltIcon } from '@patternfly/react-icons';
 import { Button } from '@patternfly/react-core';
-import { LogSnippet, Status, StatusIconAndText, BuildConfigOverviewItem } from '@console/shared';
+import { LogSnippet, Status, BuildConfigOverviewItem } from '@console/shared';
+import { StatusIconAndText } from '@console/dynamic-plugin-sdk';
 import { BuildNumberLink, BuildLogLink } from '../build';
 import { errorModal } from '../modals/error-modal';
 import { fromNow } from '../utils/datetime';
